pyproject.toml
[project]
name = "estat-mcp"
version = "0.1.0"
description = "e-Stat API MCP server"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "PyJWT[crypto]>=2.8.0",
]

[project.scripts]
estat-mcp = "src.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=1.0.0",
]
